requirements.txt
torch
transformers
sentence-transformers
langchain
tqdm
accelerate
pypdf
chromadb
langchain-community
openai
tiktoken
pypdf
pdf2image
flask
flask-cors
asgiref
python-dotenv